gcs: naze: rename some elements
