+++ ground/gcs/src/plugins/boards_naze/nazeplugin.h
@@ -2,7 +2,7 @@
  ******************************************************************************
  *
  * @file       nazeplugin.h
- * @author     Tau Labs, http://taulabs.org, Copyright (C) 2013
+ * @author     Tau Labs, http://taulabs.org, Copyright (C) 2013-2015
  *
  * @addtogroup GCSPlugins GCS Plugins
  * @{
